feat: auth check

+++ src/services/board.service/board_delete.service.js
@@ -1,13 +1,26 @@
 const pool = require('../db_connect.service.js');
 
-exports.delete_write = async (page) => {
+exports.delete_write = async (page, jwt_username, role) => {
     const connection = await pool.getConnection(async conn => conn);
 
     try {
-        await connection.execute(
-            'DELETE FROM board WHERE `index` = ?',
+        const [[{ 'username': comment_username }]] = await connection.execute(
+            'SELECT username FROM board WHERE `index` = ?',
             [page]
         )
+
+        if (jwt_username === comment_username || role === 1) {
+            await connection.execute(
+                'DELETE FROM board WHERE `index` = ?',
+                [page]
+            )
+
+            return 'succeed';
+        }
+        else {
+            return 'fail';
+        }
+
     } catch (e) {
         console.log('delete write error : ' + e);
     } finally {
